fix: Disable context menu on reaflow canvas
Previously, panning with right click would cause a context menu to
appear on right click release.
Now, the menu does not appear when right clicking on the reaflow canvas
at all, as it is not really necessary under normal circumstances.
If necessary, the context menu can still be accessed by right clicking
on other non-canvas components.